draft for group saving #112
This allows to store groups (the connected/disconnected state of the slave devices) and restore them if accessed from the correct master device.

I have very little experience in Android development, so I'd be thankfull for any hints. As far as I can tell, the major remaining issues are the following:

  - Master devices: for each group, the master device must be stored too. This would allow to either a) filter the displayed list of stored groups, by the currently-selected master device, or b) to offer all groups, but the send the correct messages to setup the selected group with the different (saved) master device.

  - option b) is related to figuring out what exact set of info & context is need to talk to the speaker devices for group setup. I'm interested in that because I'd like to add the possibility to put a button on the users' home-screen that one-click-restores a given group, without opening the full app UI.

  - Also, it would be nice to store not only which devices are connected, but also what audio channels are set on each device ("ST", "FL", "FR" ...)

(with out issue 2 and 3, the whole feature is not extremely usefull -- when the user already is on the dialog_multiroom_layout screen, a group configuration can be quickly recreated by hand, uncless a large number of devices is involved.